+++ cmd/cxplayground/main.go
@@ -1,31 +1,16 @@
 package main
 
 import (
-	"bytes"
-	"encoding/json"
 	"fmt"
-	"io"
-	"io/ioutil"
 	"net"
 	"net/http"
 	"os"
-	"runtime"
-	"time"
 
-	cxcore "github.com/skycoin/cx/cx"
 	"github.com/skycoin/cx/cxgo/actions"
 	"github.com/skycoin/cx/cxgo/api"
 
-	"github.com/skycoin/cx/cxgo/cxgo"
-	"github.com/skycoin/cx/cxgo/cxgo0"
-	"github.com/skycoin/cx/cxgo/parser"
-
 	"github.com/skycoin/cx/cmd/cxplayground/playground"
 )
-
-type SourceCode struct {
-	Code string
-}
 
 func main() {
 	host := ":5336"
@@ -42,100 +27,10 @@ func main() {
 
 	mux.Handle("/", http.FileServer(http.Dir("./dist")))
 	mux.Handle("/program/", api.NewAPI("/program", actions.PRGRM))
-	mux.HandleFunc("/eval", func(w http.ResponseWriter, r *http.Request) {
-		defer r.Body.Close()
-		var b []byte
-		var err error
-		if b, err = ioutil.ReadAll(r.Body); err != nil {
-			http.Error(w, err.Error(), 500)
-			return
-		}
-
-		var source SourceCode
-		if err := json.Unmarshal(b, &source); err != nil {
-			http.Error(w, err.Error(), 500)
-			return
-		}
-
-		if err := r.ParseForm(); err == nil {
-			fmt.Fprintf(w, "%s", Eval(source.Code+"\n"))
-		}
-	})
+	mux.HandleFunc("/eval", playground.RunProgram)
 
 	if listener, err := net.Listen("tcp", host); err == nil {
 		fmt.Println("Starting web service for CX playground on http://127.0.0.1:5336/")
 		http.Serve(listener, mux)
-	}
-}
-
-func unsafeEval(code string) (out string) {
-	var lexer *parser.Lexer
-	defer func() {
-		if r := recover(); r != nil {
-			out = fmt.Sprintf("%v", r)
-			lexer.Stop()
-		}
-	}()
-
-	// storing strings sent to standard output
-	old := os.Stdout
-	r, w, _ := os.Pipe()
-	os.Stdout = w
-
-	actions.LineNo = 0
-
-	actions.PRGRM = cxcore.MakeProgram()
-	cxgo0.PRGRM0 = actions.PRGRM
-
-	cxgo0.Parse(code)
-
-	actions.PRGRM = cxgo0.PRGRM0
-
-	lexer = parser.NewLexer(bytes.NewBufferString(code))
-	parser.Parse(lexer)
-	//yyParse(lexer)
-
-	cxgo.AddInitFunction(actions.PRGRM)
-
-	if err := actions.PRGRM.RunCompiled(0, nil); err != nil {
-		actions.PRGRM = cxcore.MakeProgram()
-		return fmt.Sprintf("%s", err)
-	}
-
-	outC := make(chan string)
-	go func() {
-		var buf bytes.Buffer
-		io.Copy(&buf, r)
-		outC <- buf.String()
-	}()
-
-	w.Close()
-	os.Stdout = old // restoring the real stdout
-	out = <-outC
-
-	actions.PRGRM = cxcore.MakeProgram()
-	return out
-}
-
-func Eval(code string) string {
-	runtime.GOMAXPROCS(2)
-	ch := make(chan string, 1)
-
-	var result string
-
-	go func() {
-		result = unsafeEval(code)
-		ch <- result
-	}()
-
-	timer := time.NewTimer(20 * time.Second)
-	defer timer.Stop()
-
-	select {
-	case <-ch:
-		return result
-	case <-timer.C:
-		actions.PRGRM = cxcore.MakeProgram()
-		return "Timed out."
 	}
 }
