+++ cmd/cxplayground/playground/playground_test.go
@@ -1,8 +1,8 @@
 package playground
 
 import (
+	"bytes"
 	"encoding/json"
-	"fmt"
 	"net/http"
 	"net/http/httptest"
 	"strings"
@@ -54,7 +54,6 @@ func TestGetExampleFileContent(t *testing.T) {
 	stubs := gostub.Stub(&getExampleFileContent, func(name string) (string, error) {
 		return mockSuccesRespone, nil
 	})
-	fmt.Println(getExampleFileContent(""))
 	defer stubs.Reset()
 	handler := http.HandlerFunc(GetExampleFileContent)
 	handler.ServeHTTP(rr, req)
@@ -67,4 +66,37 @@ func TestGetExampleFileContent(t *testing.T) {
 			rr.Body.String(), mockSuccesRespone)
 	}
 
+}
+
+func TestRunProgram(t *testing.T) {
+	mockSuccesRequest := `package main; 
+
+	func main() {
+		str.print("Hello, World!");
+	}
+	`
+	mockSuccesResponse := `Hello, World!`
+	data, _ := json.Marshal(&SourceCode{
+		Code: mockSuccesRequest,
+	})
+	req, err := http.NewRequest("POST", "/",
+		bytes.NewBuffer(data))
+	if err != nil {
+		t.Fatal(err)
+	}
+	rr := httptest.NewRecorder()
+
+	handler := http.HandlerFunc(RunProgram)
+
+	handler.ServeHTTP(rr, req)
+
+	if status := rr.Code; status != http.StatusOK {
+		t.Errorf("handler returned wrong status code: got %v want %v",
+			status, http.StatusOK)
+	}
+	if !(rr.Body.String() != mockSuccesResponse) {
+		t.Errorf("handler returned wrong Body: got:%v want:%v",
+			rr.Body.String(),
+			mockSuccesResponse)
+	}
 }
